# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/ping": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "update feedback from payment gateway",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transaction_id",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/viewmodels.TransactionResponse"
                        }
                    }
                }
            }
        },
        "/transactions": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Record a transaction",
                "parameters": [
                    {
                        "description": "Transactions",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/viewmodels.TransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/viewmodels.TransactionResponse"
                        }
                    }
                }
            }
        },
        "/transactions/{transaction_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves a transaction by transaction_id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transaction_id",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/viewmodels.TransactionResponse"
                        }
                    }
                }
            }
        },
        "/transactions/{transaction_id}/feedback": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "update feedback from payment gateway",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transaction_id",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Feedback transaction",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/viewmodels.TransactionFeedbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/viewmodels.TransactionResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "viewmodels.TransactionFeedbackRequest": {
            "type": "object",
            "properties": {
                "is_transaction_success": {
                    "type": "boolean"
                }
            }
        },
        "viewmodels.TransactionRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "card_id": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "viewmodels.TransactionResponse": {
            "type": "object",
            "properties": {
                "context": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "recommendation": {
                    "type": "string"
                },
                "request_data": {},
                "transaction_status": {
                    "type": "string"
                },
                "updated": {
                    "type": "string"
                }
            }
        }
    }
}